Krr: refactor
