+++ examples/cnn_example.py
