+++ sapsan/general/backends/fake.py
@@ -1,0 +1,20 @@
+import logging
+
+from sapsan.general.models import ExperimentBackend
+
+
+class FakeExperimentBackend(ExperimentBackend):
+    def log_parameter(self, name: str, value: str):
+        logging.info("Logging experiment '{experiment}' parameter "
+                     "{name}: {value}".format(experiment=self.name,
+                                              name=name,
+                                              value=value))
+
+    def log_artifact(self, path: str):
+        logging.info("Logging artifact {path}".format(path=path))
+
+    def log_metric(self, name: str, value: float):
+        logging.info("Logging experiment '{experiment}' metric "
+                     "{name}: {value}".format(experiment=self.name,
+                                              name=name,
+                                              value=value))
